requirements.txt
anyio==4.9.0
certifi==2025.4.26
dnspython==2.7.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
pymongo==4.13.0
python-dotenv==1.1.0
python-telegram-bot==22.1
PyYAML==6.0.2
sniffio==1.3.1
typing_extensions==4.14.0
